+++ src/enums/serviceMap.ts
@@ -1,4 +1,5 @@
 import services from './services'
+import TencentNetworkAcl from '../services/networkAcl'
 import TencentSecurityGroup from '../services/securityGroup'
 import TencentSecurityGroupRule from '../services/securityGroupRule'
 import TencentCcn from '../services/ccn'
@@ -20,6 +21,7 @@ export default {
   [services.ccn]: TencentCcn,
   [services.ccnAttachment]: TencentCcnAttachment,
   [services.customerGateway]: TencentCustomerGateway,
+  [services.networkAcl]: TencentNetworkAcl,
   [services.routeTable]: TencentRouteTable,
   [services.securityGroup]: TencentSecurityGroup,
   [services.securityGroupRule]: TencentSecurityGroupRule,