+++ src/enums/services.ts
@@ -3,6 +3,7 @@ export default {
   ccnAttachment: 'ccnAttachment',
   customerGateway: 'customerGateway',
   routeTable: 'routeTable',
+  networkAcl: 'networkAcl',
   securityGroup: 'securityGroup',
   securityGroupRule: 'securityGroupRule',
   subnet: 'subnet',